IGNITE-13151 Extracted checkpointer from DatabaseSharedManager
