+++ modules/core/src/main/java/org/apache/ignite/internal/processors/cache/persistence/GridCacheDatabaseSharedManager.java
@@ -29,9 +29,7 @@ import java.nio.file.DirectoryStream;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.nio.file.StandardOpenOption;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.Comparator;
@@ -45,21 +43,12 @@ import java.util.Set;
 import java.util.UUID;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
-import java.util.concurrent.CopyOnWriteArrayList;
-import java.util.concurrent.ExecutionException;
-import java.util.concurrent.Executor;
-import java.util.concurrent.ForkJoinPool;
-import java.util.concurrent.ForkJoinTask;
-import java.util.concurrent.ForkJoinWorkerThread;
-import java.util.concurrent.Future;
 import java.util.concurrent.LinkedBlockingQueue;
-import java.util.concurrent.RejectedExecutionException;
 import java.util.concurrent.Semaphore;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicLong;
 import java.util.concurrent.atomic.AtomicReference;
-import java.util.concurrent.atomic.LongAdder;
 import java.util.concurrent.locks.ReentrantLock;
 import java.util.concurrent.locks.ReentrantReadWriteLock;
 import java.util.function.Consumer;
@@ -76,7 +65,6 @@ import org.apache.ignite.IgniteInterruptedException;
 import org.apache.ignite.IgniteLogger;
 import org.apache.ignite.IgniteSystemProperties;
 import org.apache.ignite.cluster.ClusterNode;
-import org.apache.ignite.configuration.CheckpointWriteOrder;
 import org.apache.ignite.configuration.DataPageEvictionMode;
 import org.apache.ignite.configuration.DataRegionConfiguration;
 import org.apache.ignite.configuration.DataStorageConfiguration;
@@ -86,8 +74,6 @@ import org.apache.ignite.failure.FailureType;
 import org.apache.ignite.internal.GridKernalContext;
 import org.apache.ignite.internal.IgniteFutureTimeoutCheckedException;
 import org.apache.ignite.internal.IgniteInternalFuture;
-import org.apache.ignite.internal.IgniteInterruptedCheckedException;
-import org.apache.ignite.internal.LongJVMPauseDetector;
 import org.apache.ignite.internal.NodeStoppingException;
 import org.apache.ignite.internal.managers.discovery.GridDiscoveryManager;
 import org.apache.ignite.internal.mem.DirectMemoryProvider;
@@ -132,9 +118,7 @@ import org.apache.ignite.internal.processors.cache.persistence.checkpoint.Checkp
 import org.apache.ignite.internal.processors.cache.persistence.checkpoint.CheckpointEntryType;
 import org.apache.ignite.internal.processors.cache.persistence.checkpoint.CheckpointHistory;
 import org.apache.ignite.internal.processors.cache.persistence.checkpoint.CheckpointProgress;
-import org.apache.ignite.internal.processors.cache.persistence.checkpoint.CheckpointProgressImpl;
-import org.apache.ignite.internal.processors.cache.persistence.checkpoint.PartitionDestroyQueue;
-import org.apache.ignite.internal.processors.cache.persistence.checkpoint.PartitionDestroyRequest;
+import org.apache.ignite.internal.processors.cache.persistence.checkpoint.Checkpointer;
 import org.apache.ignite.internal.processors.cache.persistence.checkpoint.ReservationReason;
 import org.apache.ignite.internal.processors.cache.persistence.file.FileIO;
 import org.apache.ignite.internal.processors.cache.persistence.file.FileIOFactory;
@@ -142,13 +126,10 @@ import org.apache.ignite.internal.processors.cache.persistence.file.FilePageStor
 import org.apache.ignite.internal.processors.cache.persistence.file.FilePageStoreManager;
 import org.apache.ignite.internal.processors.cache.persistence.metastorage.MetaStorage;
 import org.apache.ignite.internal.processors.cache.persistence.metastorage.MetastorageLifecycleListener;
-import org.apache.ignite.internal.processors.cache.persistence.pagemem.CheckpointMetricsTracker;
 import org.apache.ignite.internal.processors.cache.persistence.pagemem.PageMemoryEx;
 import org.apache.ignite.internal.processors.cache.persistence.pagemem.PageMemoryImpl;
 import org.apache.ignite.internal.processors.cache.persistence.partstate.GroupPartitionId;
-import org.apache.ignite.internal.processors.cache.persistence.partstate.PartitionAllocationMap;
 import org.apache.ignite.internal.processors.cache.persistence.snapshot.IgniteCacheSnapshotManager;
-import org.apache.ignite.internal.processors.cache.persistence.snapshot.SnapshotOperation;
 import org.apache.ignite.internal.processors.cache.persistence.tree.io.PageIO;
 import org.apache.ignite.internal.processors.cache.persistence.tree.io.PagePartitionMetaIO;
 import org.apache.ignite.internal.processors.cache.persistence.wal.FileWALPointer;
@@ -156,31 +137,20 @@ import org.apache.ignite.internal.processors.cache.persistence.wal.crc.IgniteDat
 import org.apache.ignite.internal.processors.compress.CompressionProcessor;
 import org.apache.ignite.internal.processors.port.GridPortRecord;
 import org.apache.ignite.internal.processors.query.GridQueryProcessor;
-import org.apache.ignite.internal.util.GridConcurrentHashSet;
-import org.apache.ignite.internal.util.GridConcurrentMultiPairQueue;
 import org.apache.ignite.internal.util.GridCountDownCallback;
-import org.apache.ignite.internal.util.GridMultiCollectionWrapper;
 import org.apache.ignite.internal.util.IgniteUtils;
 import org.apache.ignite.internal.util.StripedExecutor;
 import org.apache.ignite.internal.util.TimeBag;
-import org.apache.ignite.internal.util.future.CountDownFuture;
-import org.apache.ignite.internal.util.future.GridCompoundFuture;
-import org.apache.ignite.internal.util.future.GridFinishedFuture;
-import org.apache.ignite.internal.util.future.GridFutureAdapter;
 import org.apache.ignite.internal.util.lang.GridInClosure3X;
 import org.apache.ignite.internal.util.tostring.GridToStringInclude;
 import org.apache.ignite.internal.util.typedef.F;
 import org.apache.ignite.internal.util.typedef.T2;
 import org.apache.ignite.internal.util.typedef.X;
-import org.apache.ignite.internal.util.typedef.internal.CU;
-import org.apache.ignite.internal.util.typedef.internal.LT;
 import org.apache.ignite.internal.util.typedef.internal.S;
 import org.apache.ignite.internal.util.typedef.internal.SB;
 import org.apache.ignite.internal.util.typedef.internal.U;
-import org.apache.ignite.internal.util.worker.GridWorker;
 import org.apache.ignite.lang.IgniteBiPredicate;
 import org.apache.ignite.lang.IgniteBiTuple;
-import org.apache.ignite.lang.IgniteFuture;
 import org.apache.ignite.lang.IgniteInClosure;
 import org.apache.ignite.lang.IgniteOutClosure;
 import org.apache.ignite.lang.IgnitePredicate;
@@ -190,38 +160,28 @@ import org.apache.ignite.thread.IgniteThreadPoolExecutor;
 import org.apache.ignite.transactions.TransactionState;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
-import org.jsr166.ConcurrentLinkedHashMap;
 
 import static java.nio.file.StandardOpenOption.READ;
 import static java.util.Objects.nonNull;
 import static org.apache.ignite.IgniteSystemProperties.IGNITE_CHECKPOINT_READ_LOCK_TIMEOUT;
-import static org.apache.ignite.IgniteSystemProperties.IGNITE_JVM_PAUSE_DETECTOR_THRESHOLD;
 import static org.apache.ignite.IgniteSystemProperties.IGNITE_PDS_WAL_REBALANCE_THRESHOLD;
 import static org.apache.ignite.IgniteSystemProperties.IGNITE_RECOVERY_SEMAPHORE_PERMITS;
 import static org.apache.ignite.IgniteSystemProperties.getBoolean;
 import static org.apache.ignite.IgniteSystemProperties.getInteger;
 import static org.apache.ignite.cache.CacheAtomicityMode.TRANSACTIONAL_SNAPSHOT;
-import static org.apache.ignite.failure.FailureType.CRITICAL_ERROR;
 import static org.apache.ignite.failure.FailureType.SYSTEM_CRITICAL_OPERATION_TIMEOUT;
-import static org.apache.ignite.failure.FailureType.SYSTEM_WORKER_TERMINATION;
-import static org.apache.ignite.internal.LongJVMPauseDetector.DEFAULT_JVM_PAUSE_DETECTOR_THRESHOLD;
 import static org.apache.ignite.internal.pagemem.PageIdUtils.partId;
 import static org.apache.ignite.internal.pagemem.wal.record.WALRecord.RecordType.CHECKPOINT_RECORD;
 import static org.apache.ignite.internal.pagemem.wal.record.WALRecord.RecordType.MASTER_KEY_CHANGE_RECORD;
 import static org.apache.ignite.internal.pagemem.wal.record.WALRecord.RecordType.METASTORE_DATA_RECORD;
-import static org.apache.ignite.internal.processors.cache.distributed.dht.topology.GridDhtPartitionState.LOST;
 import static org.apache.ignite.internal.processors.cache.distributed.dht.topology.GridDhtPartitionState.OWNING;
 import static org.apache.ignite.internal.processors.cache.distributed.dht.topology.GridDhtPartitionState.fromOrdinal;
 import static org.apache.ignite.internal.processors.cache.persistence.CheckpointState.FINISHED;
 import static org.apache.ignite.internal.processors.cache.persistence.CheckpointState.LOCK_RELEASED;
-import static org.apache.ignite.internal.processors.cache.persistence.CheckpointState.LOCK_TAKEN;
-import static org.apache.ignite.internal.processors.cache.persistence.CheckpointState.MARKER_STORED_TO_DISK;
-import static org.apache.ignite.internal.processors.cache.persistence.CheckpointState.PAGE_SNAPSHOT_TAKEN;
+import static org.apache.ignite.internal.processors.cache.persistence.checkpoint.Checkpointer.CHECKPOINT_LOCK_HOLD_COUNT;
+import static org.apache.ignite.internal.processors.cache.persistence.checkpoint.Checkpointer.checkpointFileName;
 import static org.apache.ignite.internal.processors.cache.persistence.file.FilePageStoreManager.TMP_FILE_MATCHER;
-import static org.apache.ignite.internal.processors.cache.persistence.tree.io.PageIO.getType;
-import static org.apache.ignite.internal.processors.cache.persistence.tree.io.PageIO.getVersion;
 import static org.apache.ignite.internal.util.IgniteUtils.checkpointBufferSize;
-import static org.apache.ignite.internal.util.IgniteUtils.hexLong;
 
 /**
  *
@@ -253,9 +213,6 @@ public class GridCacheDatabaseSharedManager extends IgniteCacheDatabaseSharedMan
      */
     private static final double PAGE_LIST_CACHE_LIMIT_THRESHOLD = 0.1;
 
-    /** Skip sync. */
-    private final boolean skipSync = getBoolean(IGNITE_PDS_CHECKPOINT_TEST_SKIP_SYNC);
-
     /** */
     private final int walRebalanceThreshold = getInteger(IGNITE_PDS_WAL_REBALANCE_THRESHOLD, 500_000);
 
@@ -264,20 +221,7 @@ public class GridCacheDatabaseSharedManager extends IgniteCacheDatabaseSharedMan
         IgniteSystemProperties.IGNITE_OVERRIDE_WRITE_THROTTLING_ENABLED);
 
     /** */
-    private final boolean skipCheckpointOnNodeStop = getBoolean(IGNITE_PDS_SKIP_CHECKPOINT_ON_NODE_STOP, false);
-
-    /** */
     private final boolean logReadLockHolders = getBoolean(IGNITE_PDS_LOG_CP_READ_LOCK_HOLDERS);
-
-    /**
-     * Starting from this number of dirty pages in checkpoint, array will be sorted with
-     * {@link Arrays#parallelSort(Comparable[])} in case of {@link CheckpointWriteOrder#SEQUENTIAL}.
-     */
-    private final int parallelSortThreshold = IgniteSystemProperties.getInteger(
-        IgniteSystemProperties.CHECKPOINT_PARALLEL_SORT_THRESHOLD, 512 * 1024);
-
-    /** Checkpoint lock hold count. */
-    private static final ThreadLocal<Integer> CHECKPOINT_LOCK_HOLD_COUNT = ThreadLocal.withInitial(() -> 0);
 
     /** Assertion enabled. */
     private static final boolean ASSERTION_ENABLED = GridCacheDatabaseSharedManager.class.desiredAssertionStatus();
@@ -303,14 +247,8 @@ public class GridCacheDatabaseSharedManager extends IgniteCacheDatabaseSharedMan
     /** Prefix for meta store records which means that checkpoint entry for some group is not applicable for WAL rebalance. */
     private static final String CHECKPOINT_INAPPLICABLE_FOR_REBALANCE = "cp-wal-rebalance-inapplicable-";
 
-    /** Timeout between partition file destroy and checkpoint to handle it. */
-    private static final long PARTITION_DESTROY_CHECKPOINT_TIMEOUT = 30 * 1000; // 30 Seconds.
-
     /** */
     private static final String CHECKPOINT_RUNNER_THREAD_PREFIX = "checkpoint-runner";
-
-    /** This number of threads will be created and used for parallel sorting. */
-    private static final int PARALLEL_SORT_THREADS = Math.min(Runtime.getRuntime().availableProcessors(), 8);
 
     /** Checkpoint thread. Needs to be volatile because it is created in exchange worker. */
     private volatile Checkpointer checkpointer;
@@ -318,17 +256,8 @@ public class GridCacheDatabaseSharedManager extends IgniteCacheDatabaseSharedMan
     /** Checkpointer thread instance. */
     private volatile IgniteThread checkpointerThread;
 
-    /** For testing only. */
-    private volatile boolean checkpointsEnabled = true;
-
-    /** For testing only. */
-    private volatile GridFutureAdapter<Void> enableChangeApplied;
-
     /** Checkpont lock. */
     ReentrantReadWriteLock checkpointLock = new ReentrantReadWriteLock();
-
-    /** */
-    private long checkpointFreq;
 
     /** */
     private CheckpointHistory cpHistory;
@@ -339,14 +268,8 @@ public class GridCacheDatabaseSharedManager extends IgniteCacheDatabaseSharedMan
     /** Checkpoint metadata directory ("cp"), contains files with checkpoint start and end */
     private File cpDir;
 
-    /** */
-    private volatile boolean printCheckpointStats = true;
-
     /** Database configuration. */
     private final DataStorageConfiguration persistenceCfg;
-
-    /** */
-    private final Collection<DbCheckpointListener> lsnrs = new CopyOnWriteArrayList<>();
 
     /** */
     private boolean stopping;
@@ -358,12 +281,6 @@ public class GridCacheDatabaseSharedManager extends IgniteCacheDatabaseSharedMan
      * not the last WAL pointer and can't be used for resumming logging.
      */
     private volatile WALPointer walTail;
-
-    /** Checkpoint runner thread pool. If null tasks are to be run in single thread */
-    @Nullable private IgniteThreadPoolExecutor asyncRunner;
-
-    /** Thread local with buffers for the checkpoint threads. Each buffer represent one page for durable memory. */
-    private ThreadLocal<ByteBuffer> threadBuf;
 
     /**
      * Lock holder for compatible folders mode. Null if lock holder was created at start node. <br>
@@ -418,9 +335,6 @@ public class GridCacheDatabaseSharedManager extends IgniteCacheDatabaseSharedMan
     private final boolean recoveryVerboseLogging =
         getBoolean(IgniteSystemProperties.IGNITE_RECOVERY_VERBOSE_LOGGING, false);
 
-    /** Pointer to a memory recovery record that should be included into the next checkpoint record. */
-    private volatile WALPointer memoryRecoveryRecordPtr;
-
     /** Page list cache limits per data region. */
     private final Map<String, AtomicLong> pageListCacheLimits = new ConcurrentHashMap<>();
 
@@ -436,8 +350,6 @@ public class GridCacheDatabaseSharedManager extends IgniteCacheDatabaseSharedMan
         persistenceCfg = cfg.getDataStorageConfiguration();
 
         assert persistenceCfg != null;
-
-        checkpointFreq = persistenceCfg.getCheckpointFrequency();
 
         truncateWalOnCpFinish = persistenceCfg.isWalHistorySizeParameterUsed()
             ? persistenceCfg.getWalHistorySize() != Integer.MAX_VALUE
@@ -505,11 +417,7 @@ public class GridCacheDatabaseSharedManager extends IgniteCacheDatabaseSharedMan
      * For test use only.
      */
     public IgniteInternalFuture<Void> enableCheckpoints(boolean enable) {
-        GridFutureAdapter<Void> fut = new GridFutureAdapter<>();
-
-        enableChangeApplied = fut;
-
-        checkpointsEnabled = enable;
+        IgniteInternalFuture<Void> fut = checkpointer.enableCheckpoints(enable);
 
         wakeupForCheckpoint("enableCheckpoints()");
 
@@ -527,37 +435,6 @@ public class GridCacheDatabaseSharedManager extends IgniteCacheDatabaseSharedMan
         );
 
         persStoreMetrics.regionMetrics(memMetricsMap.values());
-    }
-
-    /**
-     * @param allowToReplace The sign which allows to replace pages from a checkpoint by page replacer.
-     *
-     * @return holder of FullPageIds obtained from each PageMemory, overall number of dirty
-     * pages, and flag defines at least one user page became a dirty since last checkpoint.
-     */
-    private CheckpointPagesInfoHolder beginAllCheckpoints(IgniteInternalFuture<?> allowToReplace) {
-        Collection<Map.Entry<PageMemoryEx, GridMultiCollectionWrapper<FullPageId>>> res =
-            new ArrayList<>(dataRegions().size());
-
-        int pagesNum = 0;
-
-        for (DataRegion reg : dataRegions()) {
-            if (!reg.config().isPersistenceEnabled())
-                continue;
-
-            GridMultiCollectionWrapper<FullPageId> nextCpPages = ((PageMemoryEx)reg.pageMemory()).beginCheckpoint(allowToReplace);
-
-            pagesNum += nextCpPages.size();
-
-            res.add(new T2<>((PageMemoryEx)reg.pageMemory(), nextCpPages));
-        }
-
-        CheckpointProgress progress = getCheckpointer().currentProgress();
-
-        if (progress != null)
-            progress.currentCheckpointPagesCount(pagesNum);
-
-        return new CheckpointPagesInfoHolder(res, pagesNum);
     }
 
     /**
@@ -582,17 +459,6 @@ public class GridCacheDatabaseSharedManager extends IgniteCacheDatabaseSharedMan
     @Override protected void start0() throws IgniteCheckedException {
         super.start0();
 
-        threadBuf = new ThreadLocal<ByteBuffer>() {
-            /** {@inheritDoc} */
-            @Override protected ByteBuffer initialValue() {
-                ByteBuffer tmpWriteBuf = ByteBuffer.allocateDirect(pageSize());
-
-                tmpWriteBuf.order(ByteOrder.nativeOrder());
-
-                return tmpWriteBuf;
-            }
-        };
-
         snapshotMgr = cctx.snapshot();
 
         final GridKernalContext kernalCtx = cctx.kernalContext();
@@ -602,8 +468,6 @@ public class GridCacheDatabaseSharedManager extends IgniteCacheDatabaseSharedMan
 
         if (!kernalCtx.clientNode()) {
             kernalCtx.internalSubscriptionProcessor().registerDatabaseListener(new MetastorageRecoveryLifecycle());
-
-            checkpointer = new Checkpointer(cctx.igniteInstanceName(), "db-checkpoint-thread", log);
 
             cpHistory = new CheckpointHistory(kernalCtx);
 
@@ -617,6 +481,41 @@ public class GridCacheDatabaseSharedManager extends IgniteCacheDatabaseSharedMan
 
             if (!U.mkdirs(cpDir))
                 throw new IgniteCheckedException("Could not create directory for checkpoint metadata: " + cpDir);
+
+            checkpointer = new Checkpointer(
+                cctx.igniteInstanceName(),
+                "db-checkpoint-thread",
+                kernalCtx.workersRegistry(),
+                log,
+                pageSize(),
+                cctx.kernalContext().longJvmPauseDetector(),
+                kernalCtx.failure(),
+                persistenceCfg,
+                initializeCheckpointPool(),
+                snapshotMgr,
+                checkpointLock,
+                cctx.wal(),
+                cpHistory,
+                persistentStoreMetricsImpl(),
+                this::dataRegions,
+                ioFactory,
+                cpDir,
+                cctx.cache(),
+                storeMgr,
+                truncateWalOnCpFinish,
+                resolveThrottlingPolicy(),
+                this::getPageMemoryForCacheGroup,
+                new ThreadLocal<ByteBuffer>() {
+                    /** {@inheritDoc} */
+                    @Override protected ByteBuffer initialValue() {
+                        ByteBuffer tmpWriteBuf = ByteBuffer.allocateDirect(pageSize());
+
+                        tmpWriteBuf.order(ByteOrder.nativeOrder());
+
+                        return tmpWriteBuf;
+                    }
+                }
+            );
 
             final FileLockHolder preLocked = kernalCtx.pdsFolderResolver()
                     .resolveFolders()
@@ -799,7 +698,7 @@ public class GridCacheDatabaseSharedManager extends IgniteCacheDatabaseSharedMan
 
         WALPointer ptr = readPointer(file, buf);
 
-        return createCheckPointEntry(cpTs, ptr, cpId, null, CheckpointEntryType.START);
+        return checkpointer.createCheckPointEntry(cpTs, ptr, cpId, null, CheckpointEntryType.START);
     }
 
     /**
@@ -877,15 +776,45 @@ public class GridCacheDatabaseSharedManager extends IgniteCacheDatabaseSharedMan
         snapshotMgr = cctx.snapshot();
 
         if (!cctx.kernalContext().clientNode() && checkpointer == null)
-            checkpointer = new Checkpointer(cctx.igniteInstanceName(), "db-checkpoint-thread", log);
+            checkpointer = new Checkpointer(
+                cctx.igniteInstanceName(),
+                "db-checkpoint-thread",
+                cctx.kernalContext().workersRegistry(),
+                log,
+                pageSize(),
+                cctx.kernalContext().longJvmPauseDetector(),
+                cctx.kernalContext().failure(),
+                persistenceCfg,
+                initializeCheckpointPool(),
+                snapshotMgr,
+                checkpointLock,
+                cctx.wal(),
+                cpHistory,
+                persistentStoreMetricsImpl(),
+                this::dataRegions,
+                ioFactory,
+                cpDir,
+                cctx.cache(),
+                storeMgr,
+                truncateWalOnCpFinish,
+                resolveThrottlingPolicy(),
+                this::getPageMemoryForCacheGroup,
+                new ThreadLocal<ByteBuffer>() {
+                    /** {@inheritDoc} */
+                    @Override protected ByteBuffer initialValue() {
+                        ByteBuffer tmpWriteBuf = ByteBuffer.allocateDirect(pageSize());
+
+                        tmpWriteBuf.order(ByteOrder.nativeOrder());
+
+                        return tmpWriteBuf;
+                    }
+                }
+            );
 
         super.onActivate(ctx);
 
-        if (!cctx.kernalContext().clientNode()) {
-            initializeCheckpointPool();
-
+        if (!cctx.kernalContext().clientNode())
             finishRecovery();
-        }
     }
 
     /** {@inheritDoc} */
@@ -905,9 +834,9 @@ public class GridCacheDatabaseSharedManager extends IgniteCacheDatabaseSharedMan
     /**
      *
      */
-    private void initializeCheckpointPool() {
+    private IgniteThreadPoolExecutor initializeCheckpointPool() {
         if (persistenceCfg.getCheckpointThreads() > 1)
-            asyncRunner = new IgniteThreadPoolExecutor(
+            return new IgniteThreadPoolExecutor(
                 CHECKPOINT_RUNNER_THREAD_PREFIX,
                 cctx.igniteInstanceName(),
                 persistenceCfg.getCheckpointThreads(),
@@ -915,6 +844,8 @@ public class GridCacheDatabaseSharedManager extends IgniteCacheDatabaseSharedMan
                 30_000,
                 new LinkedBlockingQueue<Runnable>()
             );
+
+        return null;
     }
 
     /** {@inheritDoc} */
@@ -1111,7 +1042,7 @@ public class GridCacheDatabaseSharedManager extends IgniteCacheDatabaseSharedMan
 
             // Log MemoryRecoveryRecord to make sure that old physical records are not replayed during
             // next physical recovery.
-            memoryRecoveryRecordPtr = cctx.wal().log(new MemoryRecoveryRecord(U.currentTimeMillis()));
+            checkpointer.memoryRecoveryRecordPtr(cctx.wal().log(new MemoryRecoveryRecord(U.currentTimeMillis())));
 
             for (DatabaseLifecycleListener lsnr : getDatabaseListeners(cctx.kernalContext()))
                 lsnr.afterBinaryMemoryRestore(this, binaryState);
@@ -1140,9 +1071,12 @@ public class GridCacheDatabaseSharedManager extends IgniteCacheDatabaseSharedMan
             checkpointLock.writeLock().unlock();
         }
 
-        shutdownCheckpointer(cancel);
+        Checkpointer cp = this.checkpointer;
 
-        lsnrs.clear();
+        if (cp != null)
+            cp.shutdownCheckpointer(cancel);
+
+        this.checkpointer = null;
 
         super.onKernalStop0(cancel);
 
@@ -1393,61 +1327,6 @@ public class GridCacheDatabaseSharedManager extends IgniteCacheDatabaseSharedMan
             }
 
             return pageSize;
-        }
-    }
-
-    /**
-     * @param cancel Cancel flag.
-     */
-    @SuppressWarnings("unused")
-    private void shutdownCheckpointer(boolean cancel) {
-        Checkpointer cp = checkpointer;
-
-        if (cp != null) {
-            if (cancel)
-                cp.shutdownNow();
-            else
-                cp.cancel();
-
-            try {
-                U.join(cp);
-
-                checkpointer = null;
-            }
-            catch (IgniteInterruptedCheckedException ignore) {
-                U.warn(log, "Was interrupted while waiting for checkpointer shutdown, " +
-                    "will not wait for checkpoint to finish.");
-
-                cp.shutdownNow();
-
-                while (true) {
-                    try {
-                        U.join(cp);
-
-                        checkpointer = null;
-
-                        cp.scheduledCp.fail(new NodeStoppingException("Checkpointer is stopped during node stop."));
-
-                        break;
-                    }
-                    catch (IgniteInterruptedCheckedException ignored) {
-                        //Ignore
-                    }
-                }
-
-                Thread.currentThread().interrupt();
-            }
-        }
-
-        if (asyncRunner != null) {
-            asyncRunner.shutdownNow();
-
-            try {
-                asyncRunner.awaitTermination(2, TimeUnit.MINUTES);
-            }
-            catch (InterruptedException ignore) {
-                Thread.currentThread().interrupt();
-            }
         }
     }
 
@@ -1977,14 +1856,20 @@ public class GridCacheDatabaseSharedManager extends IgniteCacheDatabaseSharedMan
      * @param lsnr Listener.
      */
     public void addCheckpointListener(DbCheckpointListener lsnr) {
-        lsnrs.add(lsnr);
+        Checkpointer cp = this.checkpointer;
+
+        if (cp != null)
+            cp.addCheckpointListener(lsnr);
     }
 
     /**
      * @param lsnr Listener.
      */
     public void removeCheckpointListener(DbCheckpointListener lsnr) {
-        lsnrs.remove(lsnr);
+        Checkpointer cp = this.checkpointer;
+
+        if (cp != null)
+            cp.removeCheckpointListener(lsnr);
     }
 
     /**
@@ -2130,19 +2015,6 @@ public class GridCacheDatabaseSharedManager extends IgniteCacheDatabaseSharedMan
         finally {
             checkpointReadUnlock();
         }
-    }
-
-    /**
-     * @param f Consumer.
-     * @return Accumulated result for all page stores.
-     */
-    public long forAllPageStores(ToLongFunction<PageStore> f) {
-        long res = 0;
-
-        for (CacheGroupContext gctx : cctx.cache().cacheGroups())
-            res += forGroupPageStores(gctx, f);
-
-        return res;
     }
 
     /**
@@ -3073,212 +2945,11 @@ public class GridCacheDatabaseSharedManager extends IgniteCacheDatabaseSharedMan
         WALPointer walPtr,
         StripedExecutor exec
     ) throws IgniteCheckedException {
-        assert cpTs != 0;
+        assert checkpointer != null : "Checkpointer is null";
 
-        long start = System.currentTimeMillis();
-
-        Collection<DataRegion> regions = dataRegions();
-
-        CheckpointPagesInfoHolder cpPagesHolder = beginAllCheckpoints(new GridFinishedFuture<>());
-
-        // Sort and split all dirty pages set to several stripes.
-        GridConcurrentMultiPairQueue<PageMemoryEx, FullPageId> pages =
-            splitAndSortCpPagesIfNeeded(cpPagesHolder);
-
-        // Identity stores set for future fsync.
-        Collection<PageStore> updStores = new GridConcurrentHashSet<>();
-
-        AtomicInteger cpPagesCnt = new AtomicInteger();
-
-        // Shared refernce for tracking exception during write pages.
-        AtomicReference<Throwable> writePagesError = new AtomicReference<>();
-
-        for (int stripeIdx = 0; stripeIdx < exec.stripesCount(); stripeIdx++) {
-            exec.execute(stripeIdx, () -> {
-                PageStoreWriter pageStoreWriter = (fullPageId, buf, tag) -> {
-                    assert tag != PageMemoryImpl.TRY_AGAIN_TAG : "Lock is held by other thread for page " + fullPageId;
-
-                    int groupId = fullPageId.groupId();
-                    long pageId = fullPageId.pageId();
-
-                    // Write buf to page store.
-                    PageStore store = storeMgr.writeInternal(groupId, pageId, buf, tag, true);
-
-                    // Save store for future fsync.
-                    updStores.add(store);
-                };
-
-                // Local buffer for write pages.
-                ByteBuffer writePageBuf = ByteBuffer.allocateDirect(pageSize());
-
-                writePageBuf.order(ByteOrder.nativeOrder());
-
-                GridConcurrentMultiPairQueue.Result<PageMemoryEx, FullPageId> res =
-                    new GridConcurrentMultiPairQueue.Result<>();
-
-                int pagesWritten = 0;
-
-                try {
-                    while (pages.next(res)) {
-                        // Fail-fast break if some exception occurred.
-                        if (writePagesError.get() != null)
-                            break;
-
-                        PageMemoryEx pageMem = res.getKey();
-
-                        // Write page content to page store via pageStoreWriter.
-                        // Tracker is null, because no need to track checkpoint metrics on recovery.
-                        pageMem.checkpointWritePage(res.getValue(), writePageBuf, pageStoreWriter, null);
-
-                        // Add number of handled pages.
-                        pagesWritten++;
-                    }
-                }
-                catch (Throwable e) {
-                    U.error(log, "Failed to write page to pageStore: " + res);
-
-                    writePagesError.compareAndSet(null, e);
-
-                    if (e instanceof Error)
-                        throw (Error)e;
-                }
-
-                cpPagesCnt.addAndGet(pagesWritten);
-            });
-        }
-
-        // Await completion all write tasks.
-        awaitApplyComplete(exec, writePagesError);
-
-        long written = U.currentTimeMillis();
-
-        // Fsync all touched stores.
-        for (PageStore updStore : updStores)
-            updStore.sync();
-
-        long fsync = U.currentTimeMillis();
-
-        for (DataRegion memPlc : regions) {
-            if (memPlc.config().isPersistenceEnabled())
-                ((PageMemoryEx)memPlc.pageMemory()).finishCheckpoint();
-        }
-
-        ByteBuffer tmpWriteBuf = ByteBuffer.allocateDirect(pageSize());
-
-        tmpWriteBuf.order(ByteOrder.nativeOrder());
-
-        CheckpointEntry cp = prepareCheckpointEntry(
-            tmpWriteBuf,
-            cpTs,
-            cpId,
-            walPtr,
-            null,
-            CheckpointEntryType.END);
-
-        writeCheckpointEntry(tmpWriteBuf, cp, CheckpointEntryType.END);
+        checkpointer.finalizeCheckpointOnRecovery(cpTs, cpId, walPtr, exec);
 
         cctx.pageStore().finishRecover();
-
-        if (log.isInfoEnabled())
-            log.info(String.format("Checkpoint finished [cpId=%s, pages=%d, markPos=%s, " +
-                    "pagesWrite=%dms, fsync=%dms, total=%dms]",
-                cpId,
-                cpPagesCnt.get(),
-                walPtr,
-                written - start,
-                fsync - written,
-                fsync - start));
-    }
-
-    /**
-     * Prepares checkpoint entry containing WAL pointer to checkpoint record.
-     * Writes into given {@code ptrBuf} WAL pointer content.
-     *
-     * @param entryBuf Buffer to fill
-     * @param cpTs Checkpoint timestamp.
-     * @param cpId Checkpoint id.
-     * @param ptr WAL pointer containing record.
-     * @param rec Checkpoint WAL record.
-     * @param type Checkpoint type.
-     * @return Checkpoint entry.
-     */
-    private CheckpointEntry prepareCheckpointEntry(
-        ByteBuffer entryBuf,
-        long cpTs,
-        UUID cpId,
-        WALPointer ptr,
-        @Nullable CheckpointRecord rec,
-        CheckpointEntryType type
-    ) {
-        assert ptr instanceof FileWALPointer;
-
-        FileWALPointer filePtr = (FileWALPointer)ptr;
-
-        entryBuf.rewind();
-
-        entryBuf.putLong(filePtr.index());
-
-        entryBuf.putInt(filePtr.fileOffset());
-
-        entryBuf.putInt(filePtr.length());
-
-        entryBuf.flip();
-
-        return createCheckPointEntry(cpTs, ptr, cpId, rec, type);
-    }
-
-    /**
-     * Writes checkpoint entry buffer {@code entryBuf} to specified checkpoint file with 2-phase protocol.
-     *
-     * @param entryBuf Checkpoint entry buffer to write.
-     * @param cp Checkpoint entry.
-     * @param type Checkpoint entry type.
-     * @throws StorageException If failed to write checkpoint entry.
-     */
-    public void writeCheckpointEntry(ByteBuffer entryBuf, CheckpointEntry cp, CheckpointEntryType type) throws StorageException {
-        String fileName = checkpointFileName(cp, type);
-        String tmpFileName = fileName + FilePageStoreManager.TMP_SUFFIX;
-
-        try {
-            try (FileIO io = ioFactory.create(Paths.get(cpDir.getAbsolutePath(), skipSync ? fileName : tmpFileName).toFile(),
-                StandardOpenOption.CREATE_NEW, StandardOpenOption.WRITE)) {
-
-                io.writeFully(entryBuf);
-
-                entryBuf.clear();
-
-                if (!skipSync)
-                    io.force(true);
-            }
-
-            if (!skipSync)
-                Files.move(Paths.get(cpDir.getAbsolutePath(), tmpFileName), Paths.get(cpDir.getAbsolutePath(), fileName));
-        }
-        catch (IOException e) {
-            throw new StorageException("Failed to write checkpoint entry [ptr=" + cp.checkpointMark()
-                + ", cpTs=" + cp.timestamp()
-                + ", cpId=" + cp.checkpointId()
-                + ", type=" + type + "]", e);
-        }
-    }
-
-    /**
-     * @param cpTs Checkpoint timestamp.
-     * @param cpId Checkpoint ID.
-     * @param type Checkpoint type.
-     * @return Checkpoint file name.
-     */
-    private static String checkpointFileName(long cpTs, UUID cpId, CheckpointEntryType type) {
-        return cpTs + "-" + cpId + "-" + type + ".bin";
-    }
-
-    /**
-     * @param cp Checkpoint entry.
-     * @param type Checkpoint type.
-     * @return Checkpoint file name.
-     */
-    public static String checkpointFileName(CheckpointEntry cp, CheckpointEntryType type) {
-        return checkpointFileName(cp.timestamp(), cp.checkpointId(), type);
     }
 
     /**
@@ -3287,38 +2958,9 @@ public class GridCacheDatabaseSharedManager extends IgniteCacheDatabaseSharedMan
      * @param threadBuf new thread-local with buffers for the checkpoint threads.
      */
     public void setThreadBuf(final ThreadLocal<ByteBuffer> threadBuf) {
-        this.threadBuf = threadBuf;
-    }
+        assert checkpointer != null : "Checkpointer is null";
 
-    /**
-     * @param cpTs Checkpoint timestamp.
-     * @param ptr Wal pointer of checkpoint.
-     * @param cpId Checkpoint ID.
-     * @param rec Checkpoint record.
-     * @param type Checkpoint type.
-     *
-     * @return Checkpoint entry.
-     */
-    public CheckpointEntry createCheckPointEntry(
-        long cpTs,
-        WALPointer ptr,
-        UUID cpId,
-        @Nullable CheckpointRecord rec,
-        CheckpointEntryType type
-    ) {
-        assert cpTs > 0;
-        assert ptr != null;
-        assert cpId != null;
-        assert type != null;
-
-        Map<Integer, CacheState> cacheGrpStates = null;
-
-        // Do not hold groups state in-memory if there is no space in the checkpoint history to prevent possible OOM.
-        // In this case the actual group states will be readed from WAL by demand.
-        if (rec != null && cpHistory.hasSpace())
-            cacheGrpStates = rec.cacheGroupStates();
-
-        return new CheckpointEntry(cpTs, ptr, cpId, cacheGrpStates);
+        checkpointer.threadBuf(threadBuf);
     }
 
     /**
@@ -3383,1361 +3025,6 @@ public class GridCacheDatabaseSharedManager extends IgniteCacheDatabaseSharedMan
         }
 
         return null;
-    }
-
-    /**
-     * Checkpointer object is used for notification on checkpoint begin, predicate is {@link #scheduledCp}<code>.nextCpTs - now
-     * > 0 </code>. Method {@link #wakeupForCheckpoint} uses notify, {@link #waitCheckpointEvent} uses wait
-     */
-    @SuppressWarnings("NakedNotify")
-    public class Checkpointer extends GridWorker {
-        /** Checkpoint started log message format. */
-        private static final String CHECKPOINT_STARTED_LOG_FORMAT = "Checkpoint started [" +
-            "checkpointId=%s, " +
-            "startPtr=%s, " +
-            "checkpointBeforeLockTime=%dms, " +
-            "checkpointLockWait=%dms, " +
-            "checkpointListenersExecuteTime=%dms, " +
-            "checkpointLockHoldTime=%dms, " +
-            "walCpRecordFsyncDuration=%dms, " +
-            "writeCheckpointEntryDuration=%dms, " +
-            "splitAndSortCpPagesDuration=%dms, " +
-            "%s pages=%d, " +
-            "reason='%s']";
-
-        /** Temporary write buffer. */
-        private final ByteBuffer tmpWriteBuf;
-
-        /** Next scheduled checkpoint progress. */
-        private volatile CheckpointProgressImpl scheduledCp;
-
-        /** Current checkpoint. This field is updated only by checkpoint thread. */
-        private volatile CheckpointProgressImpl curCpProgress;
-
-        /** Shutdown now. */
-        private volatile boolean shutdownNow;
-
-        /** */
-        private long lastCpTs;
-
-        /** Pause detector. */
-        private final LongJVMPauseDetector pauseDetector;
-
-        /** Long JVM pause threshold. */
-        private final int longJvmPauseThreshold =
-            getInteger(IGNITE_JVM_PAUSE_DETECTOR_THRESHOLD, DEFAULT_JVM_PAUSE_DETECTOR_THRESHOLD);
-
-        /**
-         * @param gridName Grid name.
-         * @param name Thread name.
-         * @param log Logger.
-         */
-        protected Checkpointer(@Nullable String gridName, String name, IgniteLogger log) {
-            super(gridName, name, log, cctx.kernalContext().workersRegistry());
-
-            scheduledCp = new CheckpointProgressImpl(checkpointFreq);
-
-            tmpWriteBuf = ByteBuffer.allocateDirect(pageSize());
-
-            tmpWriteBuf.order(ByteOrder.nativeOrder());
-
-            pauseDetector = cctx.kernalContext().longJvmPauseDetector();
-        }
-
-        /**
-         * @return Progress of current chekpoint or {@code null}, if isn't checkpoint at this moment.
-         */
-        public CheckpointProgress currentProgress() {
-            return curCpProgress;
-        }
-
-        /** {@inheritDoc} */
-        @Override protected void body() {
-            Throwable err = null;
-
-            try {
-                while (!isCancelled()) {
-                    waitCheckpointEvent();
-
-                    if (skipCheckpointOnNodeStop && (isCancelled() || shutdownNow)) {
-                        if (log.isInfoEnabled())
-                            log.warning("Skipping last checkpoint because node is stopping.");
-
-                        return;
-                    }
-
-                    GridFutureAdapter<Void> enableChangeApplied = GridCacheDatabaseSharedManager.this.enableChangeApplied;
-
-                    if (enableChangeApplied != null) {
-                        enableChangeApplied.onDone();
-
-                        GridCacheDatabaseSharedManager.this.enableChangeApplied = null;
-                    }
-
-                    if (checkpointsEnabled)
-                        doCheckpoint();
-                    else {
-                        synchronized (this) {
-                            scheduledCp.nextCpNanos(System.nanoTime() + U.millisToNanos(checkpointFreq));
-                        }
-                    }
-                }
-
-                // Final run after the cancellation.
-                if (checkpointsEnabled && !shutdownNow)
-                    doCheckpoint();
-            }
-            catch (Throwable t) {
-                err = t;
-
-                scheduledCp.fail(t);
-
-                throw t;
-            }
-            finally {
-                if (err == null && !(stopping && isCancelled))
-                    err = new IllegalStateException("Thread is terminated unexpectedly: " + name());
-
-                if (err instanceof OutOfMemoryError)
-                    cctx.kernalContext().failure().process(new FailureContext(CRITICAL_ERROR, err));
-                else if (err != null)
-                    cctx.kernalContext().failure().process(new FailureContext(SYSTEM_WORKER_TERMINATION, err));
-
-                scheduledCp.fail(new NodeStoppingException("Node is stopping."));
-            }
-        }
-
-        /**
-         *
-         */
-        private CheckpointProgress wakeupForCheckpoint(long delayFromNow, String reason) {
-            return wakeupForCheckpoint(delayFromNow, reason, null);
-        }
-
-        /**
-         *
-         */
-        private <R> CheckpointProgress wakeupForCheckpoint(
-            long delayFromNow,
-            String reason,
-            IgniteInClosure<? super IgniteInternalFuture<R>> lsnr
-        ) {
-            if (lsnr != null) {
-                // To be sure lsnr always will be executed in checkpoint thread.
-                synchronized (this) {
-                    CheckpointProgress sched = scheduledCp;
-
-                    sched.futureFor(FINISHED).listen(lsnr);
-                }
-            }
-
-            CheckpointProgressImpl sched = scheduledCp;
-
-            long nextNanos = System.nanoTime() + U.millisToNanos(delayFromNow);
-
-            if (sched.nextCpNanos() <= nextNanos)
-                return sched;
-
-            synchronized (this) {
-                sched = scheduledCp;
-
-                if (sched.nextCpNanos() > nextNanos) {
-                    sched.reason(reason);
-
-                    sched.nextCpNanos(nextNanos);
-                }
-
-                notifyAll();
-            }
-
-            return sched;
-        }
-
-        /**
-         * @param snapshotOperation Snapshot operation.
-         */
-        public IgniteInternalFuture wakeupForSnapshotCreation(SnapshotOperation snapshotOperation) {
-            GridFutureAdapter<Object> ret;
-
-            synchronized (this) {
-                scheduledCp.nextCpNanos(System.nanoTime());
-
-                scheduledCp.reason("snapshot");
-
-                scheduledCp.nextSnapshot(true);
-
-                scheduledCp.snapshotOperation(snapshotOperation);
-
-                ret = scheduledCp.futureFor(LOCK_RELEASED);
-
-                notifyAll();
-            }
-
-            return ret;
-        }
-
-        /**
-         *
-         */
-        private void doCheckpoint() {
-            Checkpoint chp = null;
-
-            try {
-                CheckpointMetricsTracker tracker = new CheckpointMetricsTracker();
-
-                try {
-                    chp = markCheckpointBegin(tracker);
-                }
-                catch (Exception e) {
-                    if (curCpProgress != null)
-                        curCpProgress.fail(e);
-
-                    // In case of checkpoint initialization error node should be invalidated and stopped.
-                    cctx.kernalContext().failure().process(new FailureContext(FailureType.CRITICAL_ERROR, e));
-
-                    throw new IgniteException(e); // Re-throw as unchecked exception to force stopping checkpoint thread.
-                }
-
-                updateHeartbeat();
-
-                currentProgress().initCounters(chp.pagesSize);
-
-                boolean success = false;
-
-                int destroyedPartitionsCnt;
-
-                try {
-                    if (chp.hasDelta()) {
-                        // Identity stores set.
-                        ConcurrentLinkedHashMap<PageStore, LongAdder> updStores = new ConcurrentLinkedHashMap<>();
-
-                        CountDownFuture doneWriteFut = new CountDownFuture(
-                            asyncRunner == null ? 1 : persistenceCfg.getCheckpointThreads());
-
-                        tracker.onPagesWriteStart();
-
-                        final int totalPagesToWriteCnt = chp.pagesSize;
-
-                        if (asyncRunner != null) {
-                            for (int i = 0; i < persistenceCfg.getCheckpointThreads(); i++) {
-                                Runnable write = new WriteCheckpointPages(
-                                    tracker,
-                                    chp.cpPages,
-                                    updStores,
-                                    doneWriteFut,
-                                    totalPagesToWriteCnt,
-                                    this::updateHeartbeat
-                                );
-
-                                try {
-                                    asyncRunner.execute(write);
-                                }
-                                catch (RejectedExecutionException ignore) {
-                                    // Run the task synchronously.
-                                    updateHeartbeat();
-
-                                    write.run();
-                                }
-                            }
-                        }
-                        else {
-                            // Single-threaded checkpoint.
-                            updateHeartbeat();
-
-                            Runnable write = new WriteCheckpointPages(
-                                tracker,
-                                chp.cpPages,
-                                updStores,
-                                doneWriteFut,
-                                totalPagesToWriteCnt,
-                                this::updateHeartbeat);
-
-                            write.run();
-                        }
-
-                        updateHeartbeat();
-
-                        // Wait and check for errors.
-                        doneWriteFut.get();
-
-                        // Must re-check shutdown flag here because threads may have skipped some pages.
-                        // If so, we should not put finish checkpoint mark.
-                        if (shutdownNow) {
-                            chp.progress.fail(new NodeStoppingException("Node is stopping."));
-
-                            return;
-                        }
-
-                        tracker.onFsyncStart();
-
-                        if (!skipSync) {
-                            for (Map.Entry<PageStore, LongAdder> updStoreEntry : updStores.entrySet()) {
-                                if (shutdownNow) {
-                                    chp.progress.fail(new NodeStoppingException("Node is stopping."));
-
-                                    return;
-                                }
-
-                                blockingSectionBegin();
-
-                                try {
-                                    updStoreEntry.getKey().sync();
-                                }
-                                finally {
-                                    blockingSectionEnd();
-                                }
-
-                                currentProgress().updateSyncedPages(updStoreEntry.getValue().intValue());
-                            }
-                        }
-                    }
-                    else {
-                        tracker.onPagesWriteStart();
-                        tracker.onFsyncStart();
-                    }
-
-                    snapshotMgr.afterCheckpointPageWritten();
-
-                    destroyedPartitionsCnt = destroyEvictedPartitions();
-
-                    // Must mark successful checkpoint only if there are no exceptions or interrupts.
-                    success = true;
-                }
-                finally {
-                    if (success)
-                        markCheckpointEnd(chp);
-                }
-
-                tracker.onEnd();
-
-                if (chp.hasDelta() || destroyedPartitionsCnt > 0) {
-                    if (printCheckpointStats) {
-                        if (log.isInfoEnabled()) {
-                            String walSegsCoveredMsg = prepareWalSegsCoveredMsg(chp.walSegsCoveredRange);
-
-                            log.info(String.format("Checkpoint finished [cpId=%s, pages=%d, markPos=%s, " +
-                                    "walSegmentsCleared=%d, walSegmentsCovered=%s, markDuration=%dms, pagesWrite=%dms, fsync=%dms, " +
-                                    "total=%dms]",
-                                chp.cpEntry != null ? chp.cpEntry.checkpointId() : "",
-                                chp.pagesSize,
-                                chp.cpEntry != null ? chp.cpEntry.checkpointMark() : "",
-                                chp.walFilesDeleted,
-                                walSegsCoveredMsg,
-                                tracker.markDuration(),
-                                tracker.pagesWriteDuration(),
-                                tracker.fsyncDuration(),
-                                tracker.totalDuration()));
-                        }
-                    }
-                }
-
-                updateMetrics(chp, tracker);
-            }
-            catch (IgniteCheckedException e) {
-                if (chp != null)
-                    chp.progress.fail(e);
-
-                cctx.kernalContext().failure().process(new FailureContext(FailureType.CRITICAL_ERROR, e));
-            }
-        }
-
-        /**
-         * @param chp Checkpoint.
-         * @param tracker Tracker.
-         */
-        private void updateMetrics(Checkpoint chp, CheckpointMetricsTracker tracker) {
-            if (persStoreMetrics.metricsEnabled()) {
-                persStoreMetrics.onCheckpoint(
-                    tracker.lockWaitDuration(),
-                    tracker.markDuration(),
-                    tracker.pagesWriteDuration(),
-                    tracker.fsyncDuration(),
-                    tracker.totalDuration(),
-                    chp.pagesSize,
-                    tracker.dataPagesWritten(),
-                    tracker.cowPagesWritten(),
-                    forAllPageStores(PageStore::size),
-                    forAllPageStores(PageStore::getSparseSize));
-            }
-        }
-
-        /** */
-        private String prepareWalSegsCoveredMsg(IgniteBiTuple<Long, Long> walRange) {
-            String res;
-
-            long startIdx = walRange.get1();
-            long endIdx = walRange.get2();
-
-            if (endIdx < 0 || endIdx < startIdx)
-                res = "[]";
-            else if (endIdx == startIdx)
-                res = "[" + endIdx + "]";
-            else
-                res = "[" + startIdx + " - " + endIdx + "]";
-
-            return res;
-        }
-
-        /**
-         * Processes all evicted partitions scheduled for destroy.
-         *
-         * @throws IgniteCheckedException If failed.
-         *
-         * @return The number of destroyed partition files.
-         */
-        private int destroyEvictedPartitions() throws IgniteCheckedException {
-            PartitionDestroyQueue destroyQueue = curCpProgress.getDestroyQueue();
-
-            if (destroyQueue.pendingReqs().isEmpty())
-                return 0;
-
-            List<PartitionDestroyRequest> reqs = null;
-
-            for (final PartitionDestroyRequest req : destroyQueue.pendingReqs().values()) {
-                if (!req.beginDestroy())
-                    continue;
-
-                final int grpId = req.groupId();
-                final int partId = req.partitionId();
-
-                CacheGroupContext grp = cctx.cache().cacheGroup(grpId);
-
-                assert grp != null
-                    : "Cache group is not initialized [grpId=" + grpId + "]";
-                assert grp.offheap() instanceof GridCacheOffheapManager
-                    : "Destroying partition files when persistence is off " + grp.offheap();
-
-                final GridCacheOffheapManager offheap = (GridCacheOffheapManager) grp.offheap();
-
-                Runnable destroyPartTask = () -> {
-                    try {
-                        offheap.destroyPartitionStore(grpId, partId);
-
-                        req.onDone(null);
-
-                        grp.metrics().decrementInitializedLocalPartitions();
-
-                        if (log.isDebugEnabled())
-                            log.debug("Partition file has destroyed [grpId=" + grpId + ", partId=" + partId + "]");
-                    }
-                    catch (Exception e) {
-                        req.onDone(new IgniteCheckedException(
-                            "Partition file destroy has failed [grpId=" + grpId + ", partId=" + partId + "]", e));
-                    }
-                };
-
-                if (asyncRunner != null) {
-                    try {
-                        asyncRunner.execute(destroyPartTask);
-                    }
-                    catch (RejectedExecutionException ignore) {
-                        // Run the task synchronously.
-                        destroyPartTask.run();
-                    }
-                }
-                else
-                    destroyPartTask.run();
-
-                if (reqs == null)
-                    reqs = new ArrayList<>();
-
-                reqs.add(req);
-            }
-
-            if (reqs != null)
-                for (PartitionDestroyRequest req : reqs)
-                    req.waitCompleted();
-
-            destroyQueue.pendingReqs().clear();
-
-            return reqs != null ? reqs.size() : 0;
-        }
-
-        /**
-         * @param grpCtx Group context. Can be {@code null} in case of crash recovery.
-         * @param grpId Group ID.
-         * @param partId Partition ID.
-         */
-        private void schedulePartitionDestroy(@Nullable CacheGroupContext grpCtx, int grpId, int partId) {
-            synchronized (this) {
-                scheduledCp.getDestroyQueue().addDestroyRequest(grpCtx, grpId, partId);
-            }
-
-            if (log.isDebugEnabled())
-                log.debug("Partition file has been scheduled to destroy [grpId=" + grpId + ", partId=" + partId + "]");
-
-            if (grpCtx != null)
-                wakeupForCheckpoint(PARTITION_DESTROY_CHECKPOINT_TIMEOUT, "partition destroy");
-        }
-
-        /**
-         * @param grpId Group ID.
-         * @param partId Partition ID.
-         */
-        private void cancelOrWaitPartitionDestroy(int grpId, int partId) throws IgniteCheckedException {
-            PartitionDestroyRequest req;
-
-            synchronized (this) {
-                req = scheduledCp.getDestroyQueue().cancelDestroy(grpId, partId);
-            }
-
-            if (req != null)
-                req.waitCompleted();
-
-            CheckpointProgressImpl cur;
-
-            synchronized (this) {
-                cur = curCpProgress;
-
-                if (cur != null)
-                    req = cur.getDestroyQueue().cancelDestroy(grpId, partId);
-            }
-
-            if (req != null)
-                req.waitCompleted();
-
-            if (req != null && log.isDebugEnabled())
-                log.debug("Partition file destroy has cancelled [grpId=" + grpId + ", partId=" + partId + "]");
-        }
-
-        /**
-         *
-         */
-        private void waitCheckpointEvent() {
-            boolean cancel = false;
-
-            try {
-                synchronized (this) {
-                    long remaining = U.nanosToMillis(scheduledCp.nextCpNanos() - System.nanoTime());
-
-                    while (remaining > 0 && !isCancelled()) {
-                        blockingSectionBegin();
-
-                        try {
-                            wait(remaining);
-
-                            remaining = U.nanosToMillis(scheduledCp.nextCpNanos() - System.nanoTime());
-                        }
-                        finally {
-                            blockingSectionEnd();
-                        }
-                    }
-                }
-            }
-            catch (InterruptedException ignored) {
-                Thread.currentThread().interrupt();
-
-                cancel = true;
-            }
-
-            if (cancel)
-                isCancelled = true;
-        }
-
-        /**
-         *
-         */
-        @SuppressWarnings("TooBroadScope")
-        private Checkpoint markCheckpointBegin(CheckpointMetricsTracker tracker) throws IgniteCheckedException {
-            long cpTs = updateLastCheckpointTime();
-
-            CheckpointProgressImpl curr = scheduledCp;
-
-            List<DbCheckpointListener> dbLsnrs = new ArrayList<>(lsnrs);
-
-            CheckpointRecord cpRec = new CheckpointRecord(memoryRecoveryRecordPtr);
-
-            memoryRecoveryRecordPtr = null;
-
-            CheckpointEntry cp = null;
-
-            IgniteFuture snapFut = null;
-
-            CheckpointPagesInfoHolder cpPagesHolder;
-
-            int dirtyPagesCount;
-
-            boolean hasUserPages, hasPartitionsToDestroy;
-
-            DbCheckpointContextImpl ctx0 = new DbCheckpointContextImpl(curr, new PartitionAllocationMap());
-
-            internalReadLock();
-
-            try {
-                for (DbCheckpointListener lsnr : dbLsnrs)
-                    lsnr.beforeCheckpointBegin(ctx0);
-
-                ctx0.awaitPendingTasksFinished();
-            }
-            finally {
-                internalReadUnlock();
-            }
-
-            tracker.onLockWaitStart();
-
-            checkpointLock.writeLock().lock();
-
-            try {
-                updateCurrentCheckpointProgress();
-
-                assert curCpProgress == curr : "Concurrent checkpoint begin should not be happened";
-
-                tracker.onMarkStart();
-
-                // Listeners must be invoked before we write checkpoint record to WAL.
-                for (DbCheckpointListener lsnr : dbLsnrs)
-                    lsnr.onMarkCheckpointBegin(ctx0);
-
-                ctx0.awaitPendingTasksFinished();
-
-                tracker.onListenersExecuteEnd();
-
-                if (curr.nextSnapshot())
-                    snapFut = snapshotMgr.onMarkCheckPointBegin(curr.snapshotOperation(), ctx0.partitionStatMap());
-
-                fillCacheGroupState(cpRec);
-
-                // There are allowable to replace pages only after checkpoint entry was stored to disk.
-                cpPagesHolder = beginAllCheckpoints(curr.futureFor(MARKER_STORED_TO_DISK));
-
-                dirtyPagesCount = cpPagesHolder.pagesNum();
-
-                hasPartitionsToDestroy = !curr.getDestroyQueue().pendingReqs().isEmpty();
-
-                WALPointer cpPtr = null;
-
-                if (dirtyPagesCount > 0 || curr.nextSnapshot() || hasPartitionsToDestroy) {
-                    // No page updates for this checkpoint are allowed from now on.
-                    cpPtr = cctx.wal().log(cpRec);
-
-                    if (cpPtr == null)
-                        cpPtr = CheckpointStatus.NULL_PTR;
-                }
-
-                if (dirtyPagesCount > 0 || hasPartitionsToDestroy) {
-                    cp = prepareCheckpointEntry(
-                        tmpWriteBuf,
-                        cpTs,
-                        cpRec.checkpointId(),
-                        cpPtr,
-                        cpRec,
-                        CheckpointEntryType.START);
-
-                    cpHistory.addCheckpoint(cp);
-                }
-
-                curr.transitTo(PAGE_SNAPSHOT_TAKEN);
-            }
-            finally {
-                checkpointLock.writeLock().unlock();
-
-                tracker.onLockRelease();
-            }
-
-            curr.transitTo(LOCK_RELEASED);
-
-            for (DbCheckpointListener lsnr : lsnrs)
-                lsnr.onCheckpointBegin(ctx0);
-
-            if (snapFut != null) {
-                try {
-                    snapFut.get();
-                }
-                catch (IgniteException e) {
-                    U.error(log, "Failed to wait for snapshot operation initialization: " +
-                        curr.snapshotOperation(), e);
-                }
-            }
-
-            if (dirtyPagesCount > 0 || hasPartitionsToDestroy) {
-                assert cp != null;
-                assert cp.checkpointMark() != null;
-
-                tracker.onWalCpRecordFsyncStart();
-
-                // Sync log outside the checkpoint write lock.
-                cctx.wal().flush(cp.checkpointMark(), true);
-
-                tracker.onWalCpRecordFsyncEnd();
-
-                writeCheckpointEntry(tmpWriteBuf, cp, CheckpointEntryType.START);
-
-                curr.transitTo(MARKER_STORED_TO_DISK);
-
-                tracker.onSplitAndSortCpPagesStart();
-
-                GridConcurrentMultiPairQueue<PageMemoryEx, FullPageId> cpPages =
-                    splitAndSortCpPagesIfNeeded(cpPagesHolder);
-
-                tracker.onSplitAndSortCpPagesEnd();
-
-                if (printCheckpointStats && log.isInfoEnabled()) {
-                    long possibleJvmPauseDur = possibleLongJvmPauseDuration(tracker);
-
-                    if (log.isInfoEnabled())
-                        log.info(
-                            String.format(
-                                CHECKPOINT_STARTED_LOG_FORMAT,
-                                cpRec.checkpointId(),
-                                cp.checkpointMark(),
-                                tracker.beforeLockDuration(),
-                                tracker.lockWaitDuration(),
-                                tracker.listenersExecuteDuration(),
-                                tracker.lockHoldDuration(),
-                                tracker.walCpRecordFsyncDuration(),
-                                tracker.writeCheckpointEntryDuration(),
-                                tracker.splitAndSortCpPagesDuration(),
-                                possibleJvmPauseDur > 0 ? "possibleJvmPauseDuration=" + possibleJvmPauseDur + "ms," : "",
-                                dirtyPagesCount,
-                                curr.reason()
-                            )
-                        );
-                }
-
-                return new Checkpoint(cp, cpPages, curr);
-            }
-            else {
-                if (curr.nextSnapshot())
-                    cctx.wal().flush(null, true);
-
-                if (printCheckpointStats) {
-                    if (log.isInfoEnabled())
-                        LT.info(log, String.format("Skipping checkpoint (no pages were modified) [" +
-                                "checkpointBeforeLockTime=%dms, checkpointLockWait=%dms, " +
-                                "checkpointListenersExecuteTime=%dms, checkpointLockHoldTime=%dms, reason='%s']",
-                            tracker.beforeLockDuration(),
-                            tracker.lockWaitDuration(),
-                            tracker.listenersExecuteDuration(),
-                            tracker.lockHoldDuration(),
-                            curr.reason()));
-                }
-
-                return new Checkpoint(null, GridConcurrentMultiPairQueue.EMPTY, curr);
-            }
-        }
-
-        /**
-         * @param tracker Checkpoint metrics tracker.
-         * @return Duration of possible JVM pause, if it was detected, or {@code -1} otherwise.
-         */
-        private long possibleLongJvmPauseDuration(CheckpointMetricsTracker tracker) {
-            if (LongJVMPauseDetector.enabled()) {
-                if (tracker.lockWaitDuration() + tracker.lockHoldDuration() > longJvmPauseThreshold) {
-                    long now = System.currentTimeMillis();
-
-                    // We must get last wake up time before search possible pause in events map.
-                    long wakeUpTime = pauseDetector.getLastWakeUpTime();
-
-                    IgniteBiTuple<Long, Long> lastLongPause = pauseDetector.getLastLongPause();
-
-                    if (lastLongPause != null && tracker.checkpointStartTime() < lastLongPause.get1())
-                        return lastLongPause.get2();
-
-                    if (now - wakeUpTime > longJvmPauseThreshold)
-                        return now - wakeUpTime;
-                }
-            }
-
-            return -1L;
-        }
-
-        /**
-         * Take read lock for internal use.
-         */
-        private void internalReadUnlock() {
-            checkpointLock.readLock().unlock();
-
-            if (ASSERTION_ENABLED)
-                CHECKPOINT_LOCK_HOLD_COUNT.set(CHECKPOINT_LOCK_HOLD_COUNT.get() - 1);
-        }
-
-        /**
-         * Release read lock.
-         */
-        private void internalReadLock() {
-            checkpointLock.readLock().lock();
-
-            if (ASSERTION_ENABLED)
-                CHECKPOINT_LOCK_HOLD_COUNT.set(CHECKPOINT_LOCK_HOLD_COUNT.get() + 1);
-        }
-
-        /**
-         * Fill cache group state in checkpoint record.
-         *
-         * @param cpRec Checkpoint record for filling.
-         * @throws IgniteCheckedException if fail.
-         */
-        private void fillCacheGroupState(CheckpointRecord cpRec) throws IgniteCheckedException {
-            GridCompoundFuture grpHandleFut = asyncRunner == null ? null : new GridCompoundFuture();
-
-            for (CacheGroupContext grp : cctx.cache().cacheGroups()) {
-                if (grp.isLocal() || !grp.walEnabled())
-                    continue;
-
-                Runnable r = () -> {
-                    ArrayList<GridDhtLocalPartition> parts = new ArrayList<>(grp.topology().localPartitions().size());
-
-                    for (GridDhtLocalPartition part : grp.topology().currentLocalPartitions())
-                        parts.add(part);
-
-                    CacheState state = new CacheState(parts.size());
-
-                    for (GridDhtLocalPartition part : parts) {
-                        GridDhtPartitionState partState = part.state();
-
-                        if (partState == LOST)
-                            partState = OWNING;
-
-                        state.addPartitionState(
-                            part.id(),
-                            part.dataStore().fullSize(),
-                            part.updateCounter(),
-                            (byte)partState.ordinal()
-                        );
-                    }
-
-                    synchronized (cpRec) {
-                        cpRec.addCacheGroupState(grp.groupId(), state);
-                    }
-                };
-
-                if (asyncRunner == null)
-                    r.run();
-                else
-                    try {
-                        GridFutureAdapter<?> res = new GridFutureAdapter<>();
-
-                        asyncRunner.execute(U.wrapIgniteFuture(r, res));
-
-                        grpHandleFut.add(res);
-                    }
-                    catch (RejectedExecutionException e) {
-                        assert false : "Task should never be rejected by async runner";
-
-                        throw new IgniteException(e); //to protect from disabled asserts and call to failure handler
-                    }
-            }
-
-            if (grpHandleFut != null) {
-                grpHandleFut.markInitialized();
-
-                grpHandleFut.get();
-            }
-        }
-
-        /**
-         * @return Last checkpoint time.
-         */
-        private long updateLastCheckpointTime() {
-            long cpTs = System.currentTimeMillis();
-
-            // This can happen in an unlikely event of two checkpoints happening
-            // within a currentTimeMillis() granularity window.
-            if (cpTs == lastCpTs)
-                cpTs++;
-
-            lastCpTs = cpTs;
-
-            return cpTs;
-        }
-
-        /**
-         * Update current checkpoint progress by scheduled.
-         *
-         * @return Current checkpoint progress.
-         */
-        @NotNull private CheckpointProgress updateCurrentCheckpointProgress() {
-            final CheckpointProgressImpl curr;
-
-            synchronized (this) {
-                curr = scheduledCp;
-
-                curr.transitTo(LOCK_TAKEN);
-
-                if (curr.reason() == null)
-                    curr.reason("timeout");
-
-                // It is important that we assign a new progress object before checkpoint mark in page memory.
-                scheduledCp = new CheckpointProgressImpl(checkpointFreq);
-
-                curCpProgress = curr;
-            }
-            return curr;
-        }
-
-        /**
-         * @param chp Checkpoint snapshot.
-         */
-        private void markCheckpointEnd(Checkpoint chp) throws IgniteCheckedException {
-            synchronized (this) {
-                chp.progress.clearCounters();
-
-                for (DataRegion memPlc : dataRegions()) {
-                    if (!memPlc.config().isPersistenceEnabled())
-                        continue;
-
-                    ((PageMemoryEx)memPlc.pageMemory()).finishCheckpoint();
-                }
-            }
-
-            if (chp.hasDelta()) {
-                CheckpointEntry cp = prepareCheckpointEntry(
-                    tmpWriteBuf,
-                    chp.cpEntry.timestamp(),
-                    chp.cpEntry.checkpointId(),
-                    chp.cpEntry.checkpointMark(),
-                    null,
-                    CheckpointEntryType.END);
-
-                writeCheckpointEntry(tmpWriteBuf, cp, CheckpointEntryType.END);
-
-                cctx.wal().notchLastCheckpointPtr(chp.cpEntry.checkpointMark());
-            }
-
-            List<CheckpointEntry> removedFromHistory = cpHistory.onCheckpointFinished(chp, truncateWalOnCpFinish);
-
-            for (CheckpointEntry cp : removedFromHistory)
-                removeCheckpointFiles(cp);
-
-            if (chp.progress != null)
-                chp.progress.transitTo(FINISHED);
-        }
-
-        /** {@inheritDoc} */
-        @Override public void cancel() {
-            if (log.isDebugEnabled())
-                log.debug("Cancelling grid runnable: " + this);
-
-            // Do not interrupt runner thread.
-            isCancelled = true;
-
-            synchronized (this) {
-                notifyAll();
-            }
-        }
-
-        /**
-         *
-         */
-        public void shutdownNow() {
-            shutdownNow = true;
-
-            if (!isCancelled)
-                cancel();
-        }
-
-        /**
-         * Context with information about current snapshots.
-         */
-        private class DbCheckpointContextImpl implements DbCheckpointListener.Context {
-            /** Current checkpoint progress. */
-            private final CheckpointProgressImpl curr;
-
-            /** Partition map. */
-            private final PartitionAllocationMap map;
-
-            /** Pending tasks from executor. */
-            private GridCompoundFuture pendingTaskFuture;
-
-            /**
-             * @param curr Current checkpoint progress.
-             * @param map Partition map.
-             */
-            private DbCheckpointContextImpl(CheckpointProgressImpl curr, PartitionAllocationMap map) {
-                this.curr = curr;
-                this.map = map;
-                this.pendingTaskFuture = asyncRunner == null ? null : new GridCompoundFuture();
-            }
-
-            /** {@inheritDoc} */
-            @Override public CheckpointProgress progress() {
-                return curr;
-            }
-
-            /** {@inheritDoc} */
-            @Override public boolean nextSnapshot() {
-                return curr.nextSnapshot();
-            }
-
-            /** {@inheritDoc} */
-            @Override public IgniteInternalFuture<?> finishedStateFut() {
-                return curr.futureFor(FINISHED);
-            }
-
-            /** {@inheritDoc} */
-            @Override public PartitionAllocationMap partitionStatMap() {
-                return map;
-            }
-
-            /** {@inheritDoc} */
-            @Override public boolean needToSnapshot(String cacheOrGrpName) {
-                return curr.snapshotOperation().cacheGroupIds().contains(CU.cacheId(cacheOrGrpName));
-            }
-
-            /** {@inheritDoc} */
-            @Override public Executor executor() {
-                return asyncRunner == null ? null : cmd -> {
-                    try {
-                        GridFutureAdapter<?> res = new GridFutureAdapter<>();
-
-                        res.listen(fut -> updateHeartbeat());
-
-                        asyncRunner.execute(U.wrapIgniteFuture(cmd, res));
-
-                        pendingTaskFuture.add(res);
-                    }
-                    catch (RejectedExecutionException e) {
-                        assert false : "A task should never be rejected by async runner";
-                    }
-                };
-            }
-
-            /**
-             * Await all async tasks from executor was finished.
-             *
-             * @throws IgniteCheckedException if fail.
-             */
-            public void awaitPendingTasksFinished() throws IgniteCheckedException {
-                GridCompoundFuture pendingFut = this.pendingTaskFuture;
-
-                this.pendingTaskFuture = new GridCompoundFuture();
-
-                if (pendingFut != null) {
-                    pendingFut.markInitialized();
-
-                    pendingFut.get();
-                }
-            }
-        }
-    }
-
-    /**
-     * Reorders list of checkpoint pages and splits them into appropriate number of sublists according to
-     * {@link DataStorageConfiguration#getCheckpointThreads()} and
-     * {@link DataStorageConfiguration#getCheckpointWriteOrder()}.
-     *
-     * @param cpPages Checkpoint pages with overall count and user pages info.
-     */
-    private GridConcurrentMultiPairQueue<PageMemoryEx, FullPageId> splitAndSortCpPagesIfNeeded(
-        CheckpointPagesInfoHolder cpPages
-    ) throws IgniteCheckedException {
-        Set<T2<PageMemoryEx, FullPageId[]>> cpPagesPerRegion = new HashSet<>();
-
-        int realPagesArrSize = 0;
-
-        int totalPagesCnt = cpPages.pagesNum();
-
-        for (Map.Entry<PageMemoryEx, GridMultiCollectionWrapper<FullPageId>> regPages : cpPages.cpPages()) {
-            FullPageId[] pages = new FullPageId[regPages.getValue().size()];
-
-            int pagePos = 0;
-
-            for (int i = 0; i < regPages.getValue().collectionsSize(); i++) {
-                for (FullPageId page : regPages.getValue().innerCollection(i)) {
-                    if (realPagesArrSize++ == totalPagesCnt)
-                        throw new AssertionError("Incorrect estimated dirty pages number: " + totalPagesCnt);
-
-                    pages[pagePos++] = page;
-                }
-            }
-
-            // Some pages may have been already replaced.
-            if (pagePos != pages.length)
-                cpPagesPerRegion.add(new T2<>(regPages.getKey(), Arrays.copyOf(pages, pagePos)));
-            else
-                cpPagesPerRegion.add(new T2<>(regPages.getKey(), pages));
-        }
-
-        if (persistenceCfg.getCheckpointWriteOrder() == CheckpointWriteOrder.SEQUENTIAL) {
-            Comparator<FullPageId> cmp = Comparator.comparingInt(FullPageId::groupId)
-                .thenComparingLong(FullPageId::effectivePageId);
-
-            ForkJoinPool pool = null;
-
-            for (T2<PageMemoryEx, FullPageId[]> pagesPerReg : cpPagesPerRegion) {
-                if (pagesPerReg.getValue().length >= parallelSortThreshold)
-                    pool = parallelSortInIsolatedPool(pagesPerReg.get2(), cmp, pool);
-                else
-                    Arrays.sort(pagesPerReg.get2(), cmp);
-            }
-
-            if (pool != null)
-                pool.shutdown();
-        }
-
-        return new GridConcurrentMultiPairQueue<>(cpPagesPerRegion);
-    }
-
-    /**
-     * Performs parallel sort in isolated fork join pool.
-     *
-     * @param pagesArr Pages array.
-     * @param cmp Cmp.
-     *
-     * @return ForkJoinPool instance, check {@link ForkJoinTask#fork()} realization.
-     */
-    private static ForkJoinPool parallelSortInIsolatedPool(
-        FullPageId[] pagesArr,
-        Comparator<FullPageId> cmp,
-        ForkJoinPool pool
-    ) throws IgniteCheckedException {
-        ForkJoinPool.ForkJoinWorkerThreadFactory factory = new ForkJoinPool.ForkJoinWorkerThreadFactory() {
-            @Override public ForkJoinWorkerThread newThread(ForkJoinPool pool) {
-                ForkJoinWorkerThread worker = ForkJoinPool.defaultForkJoinWorkerThreadFactory.newThread(pool);
-
-                worker.setName("checkpoint-pages-sorter-" + worker.getPoolIndex());
-
-                return worker;
-            }
-        };
-
-        ForkJoinPool execPool = pool == null ?
-            new ForkJoinPool(PARALLEL_SORT_THREADS + 1, factory, null, false) : pool;
-
-        Future<?> sortTask = execPool.submit(() -> Arrays.parallelSort(pagesArr, cmp));
-
-        try {
-            sortTask.get();
-        }
-        catch (InterruptedException e) {
-            throw new IgniteInterruptedCheckedException(e);
-        }
-        catch (ExecutionException e) {
-            throw new IgniteCheckedException("Failed to perform pages array parallel sort", e.getCause());
-        }
-
-        return execPool;
-    }
-
-    /** Pages write task */
-    private class WriteCheckpointPages implements Runnable {
-        /** */
-        private final CheckpointMetricsTracker tracker;
-
-        /** Collection of page IDs to write under this task. Overall pages to write may be greater than this collection */
-        private final GridConcurrentMultiPairQueue<PageMemoryEx, FullPageId> writePageIds;
-
-        /** */
-        private final ConcurrentLinkedHashMap<PageStore, LongAdder> updStores;
-
-        /** */
-        private final CountDownFuture doneFut;
-
-        /** Total pages to write, counter may be greater than {@link #writePageIds} size */
-        private final int totalPagesToWrite;
-
-        /** */
-        private final Runnable beforePageWrite;
-
-        /**
-         * Creates task for write pages
-         *
-         * @param tracker
-         * @param writePageIds Collection of page IDs to write.
-         * @param updStores
-         * @param doneFut
-         * @param totalPagesToWrite total pages to be written under this checkpoint
-         * @param beforePageWrite Action to be performed before every page write.
-         */
-        private WriteCheckpointPages(
-            final CheckpointMetricsTracker tracker,
-            final GridConcurrentMultiPairQueue<PageMemoryEx, FullPageId> writePageIds,
-            final ConcurrentLinkedHashMap<PageStore, LongAdder> updStores,
-            final CountDownFuture doneFut,
-            final int totalPagesToWrite,
-            final Runnable beforePageWrite
-        ) {
-            this.tracker = tracker;
-            this.writePageIds = writePageIds;
-            this.updStores = updStores;
-            this.doneFut = doneFut;
-            this.totalPagesToWrite = totalPagesToWrite;
-            this.beforePageWrite = beforePageWrite;
-        }
-
-        /** {@inheritDoc} */
-        @Override public void run() {
-            snapshotMgr.beforeCheckpointPageWritten();
-
-            GridConcurrentMultiPairQueue<PageMemoryEx, FullPageId> writePageIds = this.writePageIds;
-
-            try {
-                GridConcurrentMultiPairQueue<PageMemoryEx, FullPageId> pagesToRetry = writePages(writePageIds);
-
-                if (pagesToRetry.isEmpty())
-                    doneFut.onDone();
-                else {
-                    LT.warn(log, pagesToRetry.initialSize() + " checkpoint pages were not written yet due to unsuccessful " +
-                        "page write lock acquisition and will be retried");
-
-                    while (!pagesToRetry.isEmpty())
-                        pagesToRetry = writePages(pagesToRetry);
-
-                    doneFut.onDone();
-                }
-            }
-            catch (Throwable e) {
-                doneFut.onDone(e);
-            }
-        }
-
-        /**
-         * @param writePageIds Collections of pages to write.
-         * @return pagesToRetry Pages which should be retried.
-         */
-        private GridConcurrentMultiPairQueue<PageMemoryEx, FullPageId> writePages(
-            GridConcurrentMultiPairQueue<PageMemoryEx, FullPageId> writePageIds
-        ) throws IgniteCheckedException {
-            Map<PageMemoryEx, List<FullPageId>> pagesToRetry = new HashMap<>();
-
-            CheckpointMetricsTracker tracker = persStoreMetrics.metricsEnabled() ? this.tracker : null;
-
-            PageStoreWriter pageStoreWriter = createPageStoreWriter(pagesToRetry);
-
-            ByteBuffer tmpWriteBuf = threadBuf.get();
-
-            boolean throttlingEnabled = resolveThrottlingPolicy() != PageMemoryImpl.ThrottlingPolicy.DISABLED;
-
-            GridConcurrentMultiPairQueue.Result<PageMemoryEx, FullPageId> res =
-                new GridConcurrentMultiPairQueue.Result<>();
-
-            while (writePageIds.next(res)) {
-                if (checkpointer.shutdownNow)
-                    break;
-
-                beforePageWrite.run();
-
-                FullPageId fullId = res.getValue();
-
-                PageMemoryEx pageMem = res.getKey();
-
-                snapshotMgr.beforePageWrite(fullId);
-
-                tmpWriteBuf.rewind();
-
-                pageMem.checkpointWritePage(fullId, tmpWriteBuf, pageStoreWriter, tracker);
-
-                if (throttlingEnabled) {
-                    while (pageMem.shouldThrottle()) {
-                        FullPageId cpPageId = pageMem.pullPageFromCpBuffer();
-
-                        if (cpPageId.equals(FullPageId.NULL_PAGE))
-                            break;
-
-                        snapshotMgr.beforePageWrite(cpPageId);
-
-                        tmpWriteBuf.rewind();
-
-                        pageMem.checkpointWritePage(cpPageId, tmpWriteBuf, pageStoreWriter, tracker);
-                    }
-                }
-            }
-
-            return pagesToRetry.isEmpty() ?
-                GridConcurrentMultiPairQueue.EMPTY :
-                new GridConcurrentMultiPairQueue<>(pagesToRetry);
-        }
-
-        /**
-         * Factory method for create {@link PageStoreWriter}.
-         *
-         * @param pagesToRetry List pages for retry.
-         * @return Checkpoint page write context.
-         */
-        private PageStoreWriter createPageStoreWriter(Map<PageMemoryEx, List<FullPageId>> pagesToRetry) {
-            return new PageStoreWriter() {
-                /** {@inheritDoc} */
-                @Override public void writePage(FullPageId fullPageId, ByteBuffer buf, int tag) throws IgniteCheckedException {
-                    if (tag == PageMemoryImpl.TRY_AGAIN_TAG) {
-                        PageMemoryEx pageMem = getPageMemoryForCacheGroup(fullPageId.groupId());
-
-                        pagesToRetry.computeIfAbsent(pageMem, k -> new ArrayList<>()).add(fullPageId);
-
-                        return;
-                    }
-
-                    int groupId = fullPageId.groupId();
-                    long pageId = fullPageId.pageId();
-
-                    assert getType(buf) != 0 : "Invalid state. Type is 0! pageId = " + hexLong(pageId);
-                    assert getVersion(buf) != 0 : "Invalid state. Version is 0! pageId = " + hexLong(pageId);
-
-                    if (persStoreMetrics.metricsEnabled()) {
-                        int pageType = getType(buf);
-
-                        if (PageIO.isDataPageType(pageType))
-                            tracker.onDataPageWritten();
-                    }
-
-                    getCheckpointer().currentProgress().updateWrittenPages(1);
-
-                    PageStore store = storeMgr.writeInternal(groupId, pageId, buf, tag, true);
-
-                    updStores.computeIfAbsent(store, k -> new LongAdder()).increment();
-                }
-            };
-        }
-    }
-
-    /**
-     *
-     */
-    public static class Checkpoint {
-        /** Checkpoint entry. */
-        @Nullable private final CheckpointEntry cpEntry;
-
-        /** Checkpoint pages. */
-        private final GridConcurrentMultiPairQueue<PageMemoryEx, FullPageId> cpPages;
-
-        /** */
-        private final CheckpointProgressImpl progress;
-
-        /** Number of deleted WAL files. */
-        private int walFilesDeleted;
-
-        /** WAL segments fully covered by this checkpoint. */
-        private IgniteBiTuple<Long, Long> walSegsCoveredRange;
-
-        /** */
-        private final int pagesSize;
-
-        /**
-         * @param cpEntry Checkpoint entry.
-         * @param cpPages Pages to write to the page store.
-         * @param progress Checkpoint progress status.
-         */
-        private Checkpoint(
-            @Nullable CheckpointEntry cpEntry,
-            @NotNull GridConcurrentMultiPairQueue<PageMemoryEx, FullPageId> cpPages,
-            CheckpointProgressImpl progress
-        ) {
-            this.cpEntry = cpEntry;
-            this.cpPages = cpPages;
-            this.progress = progress;
-
-            pagesSize = cpPages.initialSize();
-        }
-
-        /**
-         * @return {@code true} if this checkpoint contains at least one dirty page.
-         */
-        public boolean hasDelta() {
-            return pagesSize != 0;
-        }
-
-        /**
-         * @param walFilesDeleted Wal files deleted.
-         */
-        public void walFilesDeleted(int walFilesDeleted) {
-            this.walFilesDeleted = walFilesDeleted;
-        }
-
-        /**
-         * @param walSegsCoveredRange WAL segments fully covered by this checkpoint.
-         */
-        public void walSegsCoveredRange(final IgniteBiTuple<Long, Long> walSegsCoveredRange) {
-            this.walSegsCoveredRange = walSegsCoveredRange;
-        }
     }
 
     /**
@@ -5534,33 +3821,6 @@ public class GridCacheDatabaseSharedManager extends IgniteCacheDatabaseSharedMan
         /** */
         private CheckpointReadLockTimeoutException(String msg) {
             super(msg);
-        }
-    }
-
-    /** Current checkpoint pages information. */
-    private static class CheckpointPagesInfoHolder {
-        /** Total pages count in cp. */
-        private final int pagesNum;
-
-        /** Collection of pages per PageMemory distribution. */
-        private final Collection<Map.Entry<PageMemoryEx, GridMultiCollectionWrapper<FullPageId>>> cpPages;
-
-        /** */
-        private CheckpointPagesInfoHolder(
-            Collection<Map.Entry<PageMemoryEx, GridMultiCollectionWrapper<FullPageId>>> pages,
-            int num) {
-            cpPages = pages;
-            pagesNum = num;
-        }
-
-        /** Total pages count in cp. */
-        private int pagesNum() {
-            return pagesNum;
-        }
-
-        /** Collection of pages per PageMemory distribution. */
-        private Collection<Map.Entry<PageMemoryEx, GridMultiCollectionWrapper<FullPageId>>> cpPages() {
-            return cpPages;
         }
     }
 }
